workflow_engine/requirements.txt
# Workflow Management System - Phase 0 Foundation
# Security & Authentication
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Database & ORM
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
redis==5.0.1

# Security & Encryption
cryptography==41.0.8
python-dotenv==1.0.0
httpx==0.25.2

# Monitoring & Logging
prometheus-client==0.19.0
structlog==23.2.0
python-json-logger==2.0.7

# Validation & Schema
pydantic==2.5.0
pydantic-settings==2.1.0

# Rate Limiting & Middleware
slowapi==0.1.9
python-limiter==3.5.0

# Testing & Development
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2